function: write Shiny apps for bigKRLS objects
This is an early version of a function ultimately to be included in the bigKRLS package. To interact with the results on your machine, execute this code and then run 'shiny_bigKRLS(output).' Set 'shiny_bigKRLS(output, export=T)' if you would like to be able to use the output on a different machine or post it on a server. While we will not be able to fully anticipate everyone's research designs we will be adding features and posting vignettes to further customization. In the meantime, feedback on features likely of interest to many is most welcome!